Added some more functional specs for handling events.
Because I've updated Kiwi, beforeAll now actually only runs beforeAll (there was a bug in the old
version of Kiwi). This uncovered a bug with the onConnect() handler, which was relying on a
'connected' flag that wasn't being set.

Because all specs run under a single connection, they are now much faster.