Added libPusher framework for OSX.
Because CocoaPods doesn't support multiple platform targets right now, I've created a new
project and simply referenced everything that is needed; the libPusher source files, ZTWebSocket and JSONKit and AsyncSocket directly from the Pods directory (with ARC disabled for these).

It's a bit messy but it's the only real solution right now. Keeping it all in a separate project
makes things easier.