Binding methods should return a binding object that can be used to later remove that binding.
The problem with the last commit is that the PTPusherChannel itself wasn't retained directly
by the event dispatcher. Internally, a listener object is created for either target/action
or block based event handlers and it is these objects that retain the channel (or any other
object that is binding itself to an event).

The solution was to not have the event dispatcher retain those event listener objects directly,
but instead create a new PTPusherEventBinding object that retains the listener, and return those
all the way up the chain so that all of the relevant bind methods return the binding.

Event listeners should then keep a reference to each binding they create so they can be removed
at a later point - for an example of this, PTPusherChannel tracks any bindings it uses internally
and removes them in dealloc, to avoid it's retain cycle with the Pusher dispatcher. (#25)