Channels must bind to internal events on themselves rather than the Pusher client, otherwise they will receive internal events for all channels, when it only needs to know about internal events for itself.
A side-effect of this was that when connecting to multiple private channels, as soon as the 
first channel was authorised it would subscribe and it's subscription_succeeded would be
fired off to *all* channels, marking them as all authorised even though they haven't been.

As a result, if there was any kind of delay in authorising channels, by the time subsequent
channels started the authorisation process, they would be wrongly marked as subscribed and
wouldn't even do the authorisation request. 

This closes #26.