Make sure we always handle socket connection failures like any other type of disconnection.
When SocketRocket's connection fails, we won't always receive the
'didDisconnect' connection delegate method - sometimes we will receive
the 'didFail' method.

If we get the 'didFail' method, we'll pass that on to the Pusher delegate
whether we were already connected or not.

If we were already connected, then now it will also call the 'didDisconnect'
delegate method.

I've updated the connection monitor sample to account for this new behaviour.

This should also fix #48.