Allowing bindings to be invalidated means we can prevent them firing even if they run asynchronously.
Previously, it was possible for a binding to be removed/marked as 
invalid between the binding being told to dispatch, and the binding
actually dispatching (e.g. the block-based bindings would execute 
using dispatch_async when told to dispatch).

This caused a particular problem in PTPusherChannel, where it's internal
bindings held an unsafe unretained reference to the channel and if the
channel was dealloced between the binding being told to dispatch and
the async block call, the internal bindings would be fired with invalid
references to the channel.

__weak references to the channel would have also fixed the issue internally
but this seems like a more correct solution.