Added tests for operation failure cases (connection error and bad response).
This should fix #44 instead of using pull request #80. `responseData`
should never be nil if a successful response was returned - it should
at least be empty data.

The actual problem was caused by proceeding as if we had a response when
there was in fact a connection error - in this case it is valid that
responseData might be nil.