Deprecated the use of auto-connect and any methods that take a connectAutomatically parameter.
This behaviour was generally undesirable as you wouldn't normally
want to connect without setting a delegate first, which resulted
in extra factory methods that take the delegate or having to pass
NO and then set the delegate. It was generally confusing.

It's simpler to have just two factory methods; one with an optional
encrypted parameter, and let the user call connect when they are ready.

The deprecated methods will continue to call connect automatically if
requested but this will go away in 1.6.