We can now safely auto-reconnect if the connection fails whilst connected.
This makes it easier for clients as this can happen whenever the app 
goes into the background or the device is locked.

In the cases where there is no internet connectivity, we will still 
end up retrying at least once but this will fail and as we were not
previously connected at the point of the second failure, we will no
longer continue trying to reconnect.

We are further protected from infinite retries due to our internal
retry limit.