Ensure socket IDs still match before proceeding with subscription
It's possible that while the auth request is in progress, the connection's
socket ID changes, perhaps due to a disconnect/reconnect.

Prior to this change the subscription attempt would still proceed and
result in an "Invalid signature" error. This was mostly benign, as another
auth request was queued up with the correct socket ID, but it does
cause unnecessary errors on both the client and on Pusher, making it
harder to find legitimate issues.